requirements.txt
asgiref==3.3.4
Django==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
